requirements.txt
crewai==0.130.0
crewai-tools==0.47.1

fastapi==0.110.3
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9

pypdf>=5.0.0,<6.0.0

# CrewAI uses Pydantic v2; avoid mixing v1 + pydantic_core pins
pydantic>=2.8.2,<3.0.0

# Needed for OpenAI provider via CrewAI LLM
openai>=1.30.0

celery>=5.3.0
redis>=5.0.0